+++ types/module/module_test.go
@@ -95,7 +95,7 @@ func TestBasicManager(t *testing.T) {
 func TestAssertNoForgottenModules(t *testing.T) {
 	mockCtrl := gomock.NewController(t)
 	t.Cleanup(mockCtrl.Finish)
-	mockAppModule1 := mock.NewMockHasABCIEndblock(mockCtrl)
+	mockAppModule1 := mock.NewMockHasABCIEndBlock(mockCtrl)
 	mockAppModule3 := mock.NewMockCoreAppModule(mockCtrl)
 
 	mockAppModule1.EXPECT().Name().Times(2).Return("module1")
@@ -320,8 +320,8 @@ func TestManager_EndBlock(t *testing.T) {
 	mockCtrl := gomock.NewController(t)
 	t.Cleanup(mockCtrl.Finish)
 
-	mockAppModule1 := mock.NewMockHasABCIEndblock(mockCtrl)
-	mockAppModule2 := mock.NewMockHasABCIEndblock(mockCtrl)
+	mockAppModule1 := mock.NewMockHasABCIEndBlock(mockCtrl)
+	mockAppModule2 := mock.NewMockHasABCIEndBlock(mockCtrl)
 	mockAppModule3 := mock.NewMockAppModule(mockCtrl)
 	mockAppModule1.EXPECT().Name().Times(2).Return("module1")
 	mockAppModule2.EXPECT().Name().Times(2).Return("module2")
@@ -473,7 +473,7 @@ func TestCoreAPIManager_PreBlock(t *testing.T) {
 	mockCtrl := gomock.NewController(t)
 	t.Cleanup(mockCtrl.Finish)
 
-	mockAppModule1 := mock.NewMockCoreModuleWithPreBlock(mockCtrl)
+	mockAppModule1 := mock.NewMockCoreAppModuleWithPreBlock(mockCtrl)
 	mm := module.NewManagerFromMap(map[string]appmodule.AppModule{
 		"module1": mockAppModule1,
 		"module2": mock.NewMockCoreAppModule(mockCtrl),
@@ -482,7 +482,7 @@ func TestCoreAPIManager_PreBlock(t *testing.T) {
 	require.Equal(t, 2, len(mm.Modules))
 	require.Equal(t, 1, len(mm.OrderPreBlockers))
 
-	mockAppModule1.EXPECT().PreBlock(gomock.Any()).Times(1).Return(sdk.ResponsePreBlock{
+	mockAppModule1.EXPECT().PreBlock(gomock.Any()).Times(1).Return(&sdk.ResponsePreBlock{
 		ConsensusParamsChanged: true,
 	}, nil)
 	res, err := mm.PreBlock(sdk.Context{})
@@ -490,7 +490,7 @@ func TestCoreAPIManager_PreBlock(t *testing.T) {
 	require.True(t, res.ConsensusParamsChanged)
 
 	// test false
-	mockAppModule1.EXPECT().PreBlock(gomock.Any()).Times(1).Return(sdk.ResponsePreBlock{
+	mockAppModule1.EXPECT().PreBlock(gomock.Any()).Times(1).Return(&sdk.ResponsePreBlock{
 		ConsensusParamsChanged: false,
 	}, nil)
 	res, err = mm.PreBlock(sdk.Context{})
@@ -498,7 +498,7 @@ func TestCoreAPIManager_PreBlock(t *testing.T) {
 	require.False(t, res.ConsensusParamsChanged)
 
 	// test error
-	mockAppModule1.EXPECT().PreBlock(gomock.Any()).Times(1).Return(sdk.ResponsePreBlock{}, errors.New("some error"))
+	mockAppModule1.EXPECT().PreBlock(gomock.Any()).Times(1).Return(nil, errors.New("some error"))
 	_, err = mm.PreBlock(sdk.Context{})
 	require.EqualError(t, err, "some error")
 }