+++ x/accounts/query_server_test.go
@@ -1,6 +1,7 @@
 package accounts
 
 import (
+	"context"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -9,6 +10,7 @@ import (
 	"google.golang.org/protobuf/types/known/emptypb"
 	"google.golang.org/protobuf/types/known/wrapperspb"
 
+	bankv1beta1 "cosmossdk.io/api/cosmos/bank/v1beta1"
 	"cosmossdk.io/x/accounts/internal/implementation"
 	v1 "cosmossdk.io/x/accounts/v1"
 )
@@ -17,6 +19,9 @@ func TestQueryServer(t *testing.T) {
 	k, ctx := newKeeper(t, map[string]implementation.Account{
 		"test": TestAccount{},
 	})
+	k.queryModuleFunc = func(ctx context.Context, msg proto.Message) (proto.Message, error) {
+		return &bankv1beta1.QueryBalanceResponse{}, nil
+	}
 
 	ms := NewMsgServer(k)
 	qs := NewQueryServer(k)